added detailed ingredient route for a given recipe

+++ backend/src/main/java/com/example/recipe/model/Amount.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Amount {
-    private double value = 0;
+    private int value = 0;
     private String unit ="";
 }
